+++ addon/locales/en/components/fd-sheets-tool-bar.js
@@ -6,6 +6,8 @@ export default {
   'expand-button-title': 'Expand',
   'delete-button-title': 'Delete',
   'share-button-title': 'Copy link',
+  'dependencies-button-title': 'Exists diagrams',
+  'dependencies-panel-title': 'Diagrams where class exists',
   'print-button-title': 'Print',
   'share-button-data-content': 'Copied!',
   'link-caption': 'Link properties',